refactor(indexeDb): move indexedDb to utils
